+++ packages/paddings/src/index.ts
@@ -2,6 +2,7 @@ export * from './ansi-x923.js';
 export * from './iso10126.js';
 export * from './iso97971.js';
 export * from './none.js';
+export * from './pkcs5.js';
 export * from './pkcs7.js';
 export * from './zero.js';
 export * from './preset.js';