+++ packages/paddings/src/pkcs5.ts
@@ -1,19 +1,17 @@
 import type { PaddingComponent, PresetComponent } from '@jscrypto/core';
 import { padPkcs, unpadPkcs } from './pkcs-shared.js';
 
-export const pkcs7: PaddingComponent<'Pkcs7'> = {
+export const pkcs5: PaddingComponent<'Pkcs5'> = {
   kind: 'padding',
-  name: 'Pkcs7',
-
+  name: 'Pkcs5',
   pad: padPkcs,
-
   unpad: unpadPkcs,
 };
 
-export const pkcs7Preset: PresetComponent<'pkcs7'> = {
+export const pkcs5Preset: PresetComponent<'pkcs5'> = {
   kind: 'preset',
-  name: 'pkcs7',
+  name: 'pkcs5',
   components() {
-    return [pkcs7];
+    return [pkcs5];
   },
 };